feat(vs2022_third): C4244 fix remaining conversions in StageSPT.cpp — 1 file
Fix 8 remaining C4244 warnings by adding explicit static_cast<long> for assignments.
Lines 1072-1073: Cast double→long for local speed calculations.

No behavior changes, only explicit type conversions.

+++ SRC/Stage/StageSPT.cpp
@@ -1069,8 +1069,8 @@ int CStageSPT::GetMaxStageSpeed(double* x, double* y)
 //	XY ステージの速度を設定[StageSetLocalSpeed()] ------------------------------------------------
 void CStageSPT::SetLocalSpeed(double x, double y)
 {
-	long l_LocalSpeedX = x;
-	long l_LocalSpeedY = y;
+	long l_LocalSpeedX = static_cast<long>(x);
+	long l_LocalSpeedY = static_cast<long>(y);
 
 	m_lLocalSpeed[X] = ConvCountFrom(l_LocalSpeedX, m_lFactor[X]);
 	m_lLocalSpeed[Y] = ConvCountFrom(l_LocalSpeedY, m_lFactor[Y]);
